+++ orb-backend-status/Cargo.toml
@@ -14,6 +14,7 @@ rust-version.workspace = true
 bon.workspace = true
 chrono = { workspace = true, features = ["serde"] }
 color-eyre.workspace = true
+derive_more = { workspace = true, default-features = false, features = ["from"] }
 flume.workspace = true
 futures-util = { version = "0.3.31" }
 eyre.workspace = true
