+++ orb-backend-status/src/backend/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod types;
 mod uptime;
 
+pub mod client;
 pub mod os_version;
-pub mod status;
+pub mod status_req_builder;
