+++ orb-backend-status/src/lib.rs
@@ -1,13 +1,14 @@
 pub mod backend;
 pub mod collectors;
 pub mod dbus;
+pub mod oes_cache;
 pub mod oes_flusher;
 #[allow(dead_code)]
 pub(crate) mod oes_reroute;
 pub mod sender;
 
-use crate::{oes_reroute::OesReroute, sender::BackendSender};
-use backend::status::StatusClient;
+use crate::{oes_cache::OesEventCache, oes_reroute::OesReroute, sender::BackendSender};
+use backend::client::StatusClient;
 use collectors::{
     connectivity::{self, GlobalConnectivity},
     core_signups, front_als, hardware_states, net_stats, oes,
@@ -38,8 +39,6 @@ pub async fn program(
     orb_jabil_id: OrbJabilId,
     net_stats_poll_interval: Duration,
     sender_interval: Duration,
-    sender_min_backoff: Duration,
-    sender_max_backoff: Duration,
     req_timeout: Duration,
     req_min_retry_interval: Duration,
     req_max_retry_interval: Duration,
@@ -49,33 +48,29 @@ pub async fn program(
     info!("Starting backend-status, endpoint: {endpoint}, orb_id: {orb_id}, orb_name: {orb_name}, orb_jabil_id: {orb_jabil_id}");
 
     let backend_status_impl = BackendStatusImpl::new();
+    let oes_cache = OesEventCache::default();
 
     setup_dbus(&dbus, backend_status_impl.clone()).await?;
 
     let token_receiver =
         TokenWatcher::spawn(dbus.clone(), shutdown_token.clone()).await;
 
-    let oes_endpoint = endpoint.clone();
-    let oes_orb_id = orb_id.clone();
+    // Build unified zenorb context and single receiver
+    let (connectivity_tx, connectivity_receiver) =
+        watch::channel(GlobalConnectivity::NotConnected);
 
-    let status_client = StatusClient::new(
-        endpoint,
-        orb_os_version,
-        orb_id,
-        orb_name,
-        orb_jabil_id,
-        req_timeout,
-        req_min_retry_interval,
-        req_max_retry_interval,
-    )
-    .await?;
-
-    let sender = BackendSender::new(
-        status_client,
-        sender_interval,
-        sender_min_backoff,
-        sender_max_backoff,
-    );
+    let status_client = StatusClient::builder()
+        .orb_id(orb_id)
+        .orb_name(orb_name)
+        .jabil_id(orb_jabil_id)
+        .orb_os_version(orb_os_version)
+        .endpoint(endpoint)
+        .req_timeout(req_timeout)
+        .min_req_retry_interval(req_min_retry_interval)
+        .max_req_retry_interval(req_max_retry_interval)
+        .attest_token_rx(token_receiver)
+        .connectivity_rx(connectivity_receiver.clone())
+        .build();
 
     // Spawn non-zenorb collectors
     let mut tasks: Vec<JoinHandle<()>> = vec![];
@@ -99,10 +94,6 @@ pub async fn program(
         shutdown_token.clone(),
     ));
 
-    // Build unified zenorb context and single receiver
-    let (connectivity_tx, connectivity_receiver) =
-        watch::channel(GlobalConnectivity::NotConnected);
-
     let (oes_tx, oes_rx) = flume::unbounded();
 
     let zenorb_ctx = ZenorbCtx {
@@ -112,6 +103,7 @@ pub async fn program(
         front_als: Arc::new(tokio::sync::Mutex::new(None)),
         oes_tx,
         oes_throttle: Arc::new(std::sync::Mutex::new(HashMap::new())),
+        oes_cache: oes_cache.clone(),
     };
 
     let zenorb_tasks = zsession
@@ -140,6 +132,17 @@ pub async fn program(
         .run()
         .await?;
 
+    tasks.push(tokio::spawn(oes_flusher::run_oes_flush_loop(
+        oes_rx,
+        status_client.clone(),
+        shutdown_token.clone(),
+    )));
+
+    let sender = BackendSender::new(status_client.clone(), oes_cache, sender_interval);
+    sender
+        .run_loop(backend_status_impl, shutdown_token.clone())
+        .await;
+
     // Spawn a single shutdown task for all zenorb subscribers
     let shutdown = shutdown_token.clone();
     tasks.push(tokio::spawn(async move {
@@ -148,25 +151,6 @@ pub async fn program(
             task.abort();
         }
     }));
-
-    // Spawn OES flush loop
-    tasks.push(tokio::spawn(oes_flusher::run_oes_flush_loop(
-        oes_rx,
-        oes_endpoint,
-        oes_orb_id,
-        token_receiver.clone(),
-        connectivity_receiver.clone(),
-        shutdown_token.clone(),
-    )));
-
-    sender
-        .run_loop(
-            backend_status_impl,
-            token_receiver,
-            connectivity_receiver,
-            shutdown_token.clone(),
-        )
-        .await;
 
     for task in tasks {
         task.abort();
