+++ orb-backend-status/src/main.rs
@@ -64,8 +64,6 @@ async fn main() -> Result<()> {
         .procfs("/proc")
         .net_stats_poll_interval(Duration::from_secs(30))
         .sender_interval(Duration::from_secs(30))
-        .sender_min_backoff(Duration::from_secs(1))
-        .sender_max_backoff(Duration::from_secs(30))
         .req_timeout(Duration::from_secs(2))
         .req_min_retry_interval(Duration::from_millis(100))
         .req_max_retry_interval(Duration::from_secs(500))
