+++ orb-backend-status/src/oes_flusher.rs
@@ -1,48 +1,21 @@
+use crate::backend::client::{self, StatusClient};
 use crate::backend::types::OrbStatusApiV2;
-use crate::collectors::connectivity::GlobalConnectivity;
 use crate::collectors::oes::Event;
-use chrono::Utc;
-use orb_info::OrbId;
-use reqwest::Url;
 use std::time::Duration;
-use tokio::sync::watch;
-use tokio::time::{self, Instant};
+use tokio::time::{self};
 use tokio_util::sync::CancellationToken;
 use tracing::{debug, error, warn};
 
 pub async fn run_oes_flush_loop(
     oes_rx: flume::Receiver<Event>,
-    endpoint: Url,
-    orb_id: OrbId,
-    token_receiver: watch::Receiver<String>,
-    connectivity_receiver: watch::Receiver<GlobalConnectivity>,
+    client: StatusClient,
     shutdown_token: CancellationToken,
 ) {
-    let client = reqwest::Client::builder()
-        .timeout(Duration::from_secs(5))
-        .build()
-        .expect("failed to build OES reqwest client");
-
     let mut buffer: Vec<Event> = Vec::new();
-    let mut last_flush = Instant::now() - Duration::from_secs(1);
-    let mut backoff = Duration::from_secs(1);
     let mut interval = time::interval(Duration::from_secs(1));
-    let token_backoff = Duration::from_secs(5);
     interval.set_missed_tick_behavior(time::MissedTickBehavior::Skip);
 
     loop {
-        let token = token_receiver.borrow().clone();
-        if token.is_empty() {
-            warn!(
-                "oes_flusher could not get auth token. waiting {}s and trying again",
-                token_backoff.as_secs()
-            );
-
-            time::sleep(token_backoff).await;
-
-            continue;
-        }
-
         tokio::select! {
             _ = shutdown_token.cancelled() => {
                 if !buffer.is_empty() {
@@ -50,11 +23,9 @@ pub async fn run_oes_flush_loop(
                         count = buffer.len(),
                         "Shutdown: attempting final OES flush",
                     );
+
                     if let Err(e) = flush_events(
                         &client,
-                        &endpoint,
-                        &orb_id,
-                        &token,
                         &buffer,
                     ).await {
                         warn!("Final OES flush failed: {e}");
@@ -70,6 +41,7 @@ pub async fn run_oes_flush_loop(
                         buffer.push(event);
                         drain_available(&oes_rx, &mut buffer);
                     }
+
                     Err(_) => {
                         debug!("OES channel closed, exiting flush loop");
 
@@ -81,17 +53,7 @@ pub async fn run_oes_flush_loop(
             _ = interval.tick() => {}
         }
 
-        maybe_flush(
-            &client,
-            &endpoint,
-            &orb_id,
-            &token,
-            &mut buffer,
-            &mut last_flush,
-            &mut backoff,
-            &connectivity_receiver,
-        )
-        .await;
+        maybe_flush(&client, &mut buffer).await;
     }
 }
 
@@ -101,82 +63,55 @@ fn drain_available(rx: &flume::Receiver<Event>, buffer: &mut Vec<Event>) {
     }
 }
 
-const MAX_BACKOFF: Duration = Duration::from_secs(60);
 const MAX_BATCH_EVENTS: usize = 100;
 
 #[allow(clippy::too_many_arguments)]
-async fn maybe_flush(
-    client: &reqwest::Client,
-    endpoint: &Url,
-    orb_id: &OrbId,
-    token: &str,
-    buffer: &mut Vec<Event>,
-    last_flush: &mut Instant,
-    backoff: &mut Duration,
-    connectivity_receiver: &watch::Receiver<GlobalConnectivity>,
-) {
+async fn maybe_flush(client: &StatusClient, buffer: &mut Vec<Event>) {
     if buffer.is_empty() {
-        return;
-    }
-
-    if last_flush.elapsed() < *backoff {
-        return;
-    }
-
-    if !connectivity_receiver.borrow().is_connected() {
-        debug!(count = buffer.len(), "Orb offline, skipping OES flush");
-
         return;
     }
 
     let batch_size = buffer.len().min(MAX_BATCH_EVENTS);
     let batch = &buffer[..batch_size];
 
-    match flush_events(client, endpoint, orb_id, token, batch).await {
-        Ok(()) => {
-            debug!(count = batch_size, "OES flush successful");
-            buffer.drain(..batch_size);
-            *last_flush = Instant::now();
-            *backoff = Duration::from_secs(1);
+    match flush_events(client, batch).await {
+        Ok(sent) => {
+            if sent {
+                debug!(count = batch_size, "OES flush successful");
+                buffer.drain(..batch_size);
+            }
         }
+
         Err(e) => {
             error!(
                 count = buffer.len(),
                 "OES flush failed, events remain buffered: {e}",
             );
-            *last_flush = Instant::now();
-            *backoff = (*backoff * 2).min(MAX_BACKOFF);
         }
     }
 }
 
-async fn flush_events(
-    client: &reqwest::Client,
-    endpoint: &Url,
-    orb_id: &OrbId,
-    token: &str,
-    events: &[Event],
-) -> eyre::Result<()> {
-    let request = OrbStatusApiV2 {
-        orb_id: Some(orb_id.to_string()),
+async fn flush_events(client: &StatusClient, events: &[Event]) -> eyre::Result<bool> {
+    let req = OrbStatusApiV2 {
         oes: Some(events.to_vec()),
-        timestamp: Utc::now(),
         ..Default::default()
     };
 
-    let response = client
-        .post(endpoint.clone())
-        .json(&request)
-        .basic_auth(orb_id.to_string(), Some(token))
-        .send()
-        .await?;
+    let res = match client.req(req).await {
+        Err(client::Err::MissingAttestToken | client::Err::NoConnectivity) => {
+            return Ok(false);
+        }
 
-    let status = response.status();
+        Err(client::Err::Other(e)) => return Err(e),
+
+        Ok(res) => res,
+    };
+
+    let status = res.status();
     if !status.is_success() {
-        let body = response.text().await.unwrap_or_default();
-
+        let body = res.text().await.unwrap_or_default();
         return Err(eyre::eyre!("OES flush error: {status} - {body}"));
     }
 
-    Ok(())
+    Ok(true)
 }
