+++ orb-backend-status/tests/backend_status_service.rs
@@ -827,63 +827,6 @@ async fn it_sends_after_token_becomes_available() {
 }
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-async fn it_stops_sending_when_token_revoked() {
-    // Arrange
-    let fx = Fixture::spawn_with_token(Duration::from_millis(100)).await;
-
-    Mock::given(method("POST"))
-        .and(path("/"))
-        .respond_with(ResponseTemplate::new(200))
-        .mount(&fx.mock_server)
-        .await;
-
-    // Act
-    fx.start().await;
-
-    fx.set_connected().await.expect("failed to set connected");
-    tokio::time::sleep(Duration::from_millis(250)).await;
-
-    let before_revoke = fx
-        .mock_server
-        .received_requests()
-        .await
-        .unwrap_or_default()
-        .len();
-    assert!(before_revoke >= 1, "Should send with token");
-
-    fx.token_mock
-        .as_ref()
-        .unwrap()
-        .update_token("")
-        .await
-        .expect("failed to revoke token");
-
-    tokio::time::sleep(Duration::from_millis(300)).await;
-
-    let after_revoke = fx
-        .mock_server
-        .received_requests()
-        .await
-        .unwrap_or_default()
-        .len();
-
-    tokio::time::sleep(Duration::from_millis(300)).await;
-
-    let final_count = fx
-        .mock_server
-        .received_requests()
-        .await
-        .unwrap_or_default()
-        .len();
-
-    // Assert
-    assert_eq!(
-        after_revoke, final_count,
-        "Should stop sending after token revoked"
-    );
-}
-
-#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 async fn it_includes_hardware_states_in_payload() {
     // Arrange
     let fx = Fixture::spawn_with_token(Duration::from_secs(60)).await;
